+++ Mytube/src/main/java/com/flab/Mytube/movie/request/InsertMovieRequest.java
@@ -1,4 +1,4 @@
-package com.flab.Mytube.dto.movie.request;
+package com.flab.Mytube.movie.request;
 
 import lombok.Getter;
 import lombok.Setter;
