+++ Mytube/src/main/java/com/flab/Mytube/movie/response/InsertMovieResponse.java
@@ -1,15 +1,16 @@
-package com.flab.Mytube.dto.movie.response;
+package com.flab.Mytube.movie.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+import lombok.Setter;
 
 import java.math.BigInteger;
 
 @Getter
+@Setter
 @AllArgsConstructor
-public class InsertPostResponse {
-    private BigInteger ID;
+public class InsertMovieResponse {
+    private long ID;
     private int code;
     private String message;
 }
