+++ Mytube/src/main/java/com/flab/Mytube/movie/response/InsertPostResponse.java
@@ -1,13 +1,14 @@
-package com.flab.Mytube.dto.movie.response;
+package com.flab.Mytube.movie.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
 import java.math.BigInteger;
 
 @Getter
 @AllArgsConstructor
-public class JoinChatResponse {
+public class InsertPostResponse {
     private BigInteger ID;
     private int code;
     private String message;
