+++ Mytube/src/main/java/com/flab/Mytube/movie/response/JoinChatResponse.java
@@ -1,16 +1,14 @@
-package com.flab.Mytube.dto.movie.response;
+package com.flab.Mytube.movie.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.Setter;
 
 import java.math.BigInteger;
 
 @Getter
-@Setter
 @AllArgsConstructor
-public class InsertMovieResponse {
-    private long ID;
+public class JoinChatResponse {
+    private BigInteger ID;
     private int code;
     private String message;
 }
