+++ Mytube/src/main/java/com/flab/Mytube/movie/response/Response.java
@@ -1,4 +1,4 @@
-package com.flab.Mytube.dto.movie.response;
+package com.flab.Mytube.movie.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
